BaseColliderStateテンプレート追加・BaseStateテンプレート追加
今までBaseStateで当たり判定などを取得していたが
カメラなどで使用するため、当たり判定取得処理をBaseColliderStateにStateの更新処理などをBaseStateに分けました。
