+++ Contest/Assets/Object/Base/State/BaseColliderState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5717232e886eb3945ae113f18d6222bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
