+++ Contest/Assets/Object/ThrowObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 72c29f18237314744913af58cc09b51d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
