+++ Contest/Assets/Object/ThrowObject/ThrowObjectState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f64eccf2513b56d48a8a8de6e5bc4c8a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
